# be/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for uploading and managing images with MinIO and AntD frontend.",
        "title": "AntD Minio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/csv/{name}": {
            "get": {
                "description": "Downloads a CSV file from MinIO storage",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "CSV"
                ],
                "summary": "Download a CSV file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the CSV file (without extension)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates an empty CSV file and uploads it to MinIO storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CSV"
                ],
                "summary": "Create and upload a CSV file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the CSV file (without extension)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/image-collections": {
            "get": {
                "description": "Retrieve metadata for all uploaded image collections",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ImageCollections"
                ],
                "summary": "Get all image collections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.ImageCollectionInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Upload multiple images along with metadata (name, description, date)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ImageCollections"
                ],
                "summary": "Upload a new image collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metadata JSON for the image collection (name, description, date)",
                        "name": "info",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Multiple image files",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/image-collections/{name}": {
            "delete": {
                "description": "Remove all files in the specified image collection",
                "tags": [
                    "ImageCollections"
                ],
                "summary": "Delete an image collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name to delete",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/image-collections/{name}/images": {
            "get": {
                "description": "Retrieve all image names from a specific collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ImageCollections"
                ],
                "summary": "Get all images in a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.ImageCollectionInfo": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}